Struct nicehash::Client

pub struct Client { /* fields omitted */ }

nicehash.com API client.

Methods

impl Client

Public API methods.

fn global_stats_current(&self, location: Option<Location>) -> Result<GlobalStats>

Gets current global stats.

fn global_stats_24h(&self) -> Result<GlobalStats>

Gets global stats for the last 24h.

fn get_orders(&self, location: Location, algorithm: Algorithm) -> Result<Vec<Order>>

Gets all orders for the given algorithm and location.

fn get_buy_info(&self) -> Result<BuyInfo>

Gets needed information for buying hashing power.

impl Client

Private API methods.

fn get_my_orders<S: AsRef<str>>(&self, api_id: u64, api_key: S, location: Location, algorithm: Algorithm) -> Result<Vec<Order>>

Gets orders for the current user.

fn create_order<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, order: NewOrder, pool: PoolInfo, code: Option<u32>) -> Result<u64>

Creates a new order.

fn refill_order<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, algorithm: Algorithm, order_id: u64, amount: f64) -> Result<()>

Refills the given order with the given amount.

fn remove_order<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, algorithm: Algorithm, order_id: u64) -> Result<()>

Removes the given order.

fn set_order_price<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, algorithm: Algorithm, order_id: u64, price: f64) -> Result<()>

Sets the price to the given order.

fn decrease_order_price<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, algorithm: Algorithm, order_id: u64) -> Result<f64>

Decrease the price to the given order.

The decrease can only be done each 600 seconds (10 minutes). This number might change in the future and can be obtained with the get_buy_info() call. The price decrease is always done in finite amounts that are different for each algorithm and can also be obtained in that call.

fn set_order_speed_limit<K: AsRef<str>>(&self, api_id: u64, api_key: K, location: Location, algorithm: Algorithm, order_id: u64, speed_limit: Option<f64>) -> Result<()>

Sets the speed limit for the given order.

The decrease can only be done each 600 seconds (10 minutes). This number might change in the future and can be obtained with the get_buy_info() call.

fn get_balance<K: AsRef<str>>(&self, api_id: u64, api_key: K) -> Result<Balance>

Gets the balance of the given account.

impl Client

fn new() -> Result<Client>

Creates a new API client with default settings.

It will also call the API to check that the connection works and to store the remote API version.

fn from_hyper_client(hyper_client: Client) -> Result<Client>

Creates a new API client from a Hyper client.

fn get_api_version(&self) -> &Version

Gets the version of the remote API.

This only returns the version stored in the client, it will not update the version if remote API is updated while the client has been created.

fn update_api_version(&mut self) -> Result<&Version>

Updates the API version of this Client.

This will check the current version of the remote server and store it in the Client object for later use.